Return arrays rather than objects from all hooks
