+++ auth/helpers/index.ts
@@ -1,0 +1,10 @@
+import { auth, FirebaseError } from 'firebase';
+
+export const transformError = (error: auth.Error): FirebaseError => {
+  return {
+    message: error.message,
+    stack: '',
+    name: '',
+    code: error.code,
+  };
+};
