+++ database/index.ts
@@ -1,5 +1,14 @@
-export { default as useList, ListHook } from './useList';
-export { default as useListKeys, ListKeysHook } from './useListKeys';
-export { default as useListVals, ListValsHook } from './useListVals';
-export { default as useObject, ObjectHook } from './useObject';
-export { default as useObjectVal, ObjectValHook } from './useObjectVal';
+export {
+  useList,
+  useListKeys,
+  useListVals,
+  ListHook,
+  ListKeysHook,
+  ListValsHook,
+} from './useList';
+export {
+  useObject,
+  useObjectVal,
+  ObjectHook,
+  ObjectValHook,
+} from './useObject';
