+++ firestore/index.ts
@@ -1,14 +1,24 @@
-export { default as useCollection, CollectionHook } from './useCollection';
-export { default as useCollectionOnce } from './useCollectionOnce';
 export {
-  default as useCollectionData,
+  useCollection,
+  useCollectionData,
+  CollectionHook,
   CollectionDataHook,
-} from './useCollectionData';
-export { default as useCollectionDataOnce } from './useCollectionDataOnce';
-export { default as useDocument, DocumentHook } from './useDocument';
-export { default as useDocumentOnce } from './useDocumentOnce';
+} from './useCollection';
 export {
-  default as useDocumentData,
+  useCollectionOnce,
+  useCollectionDataOnce,
+  CollectionOnceHook,
+  CollectionDataOnceHook,
+} from './useCollectionOnce';
+export {
+  useDocument,
+  useDocumentData,
+  DocumentHook,
   DocumentDataHook,
-} from './useDocumentData';
-export { default as useDocumentDataOnce } from './useDocumentDataOnce';
+} from './useDocument';
+export {
+  useDocumentOnce,
+  useDocumentDataOnce,
+  DocumentOnceHook,
+  DocumentDataOnceHook,
+} from './useDocumentOnce';
