+++ firestore/useDocumentOnce.ts
@@ -1,12 +1,18 @@
-import { firestore } from 'firebase';
+import { firestore, FirebaseError } from 'firebase';
 import { useEffect } from 'react';
-import { useIsEqualRef, useLoadingValue } from '../util';
-import { DocumentHook } from './useDocument';
+import { snapshotToData } from './helpers';
+import { LoadingHook, useIsEqualRef, useLoadingValue } from '../util';
 
-export default (
+export type DocumentOnceHook = LoadingHook<
+  firestore.DocumentSnapshot,
+  FirebaseError
+>;
+export type DocumentDataOnceHook<T> = LoadingHook<T, FirebaseError>;
+
+export const useDocumentOnce = (
   docRef?: firestore.DocumentReference | null,
   options?: firestore.GetOptions
-): DocumentHook => {
+): DocumentOnceHook => {
   const { error, loading, reset, setError, setValue, value } = useLoadingValue<
     firestore.DocumentSnapshot
   >();
@@ -26,9 +32,22 @@ export default (
     [ref.current]
   );
 
-  return {
-    error,
+  return [value, loading, error];
+};
+
+export const useDocumentDataOnce = <T>(
+  docRef?: firestore.DocumentReference | null,
+  options?: {
+    getOptions?: firestore.GetOptions;
+    idField?: string;
+  }
+): DocumentDataOnceHook<T> => {
+  const idField = options ? options.idField : undefined;
+  const getOptions = options ? options.getOptions : undefined;
+  const [value, loading, error] = useDocumentOnce(docRef, getOptions);
+  return [
+    (value ? snapshotToData(value, idField) : undefined) as T,
     loading,
-    value,
-  };
+    error,
+  ];
 };
