+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-firebase-hooks",
-  "version": "1.2.1",
+  "version": "2.0.0-alpha.2",
   "description": "React Hooks for Firebase",
   "author": "CS Frequency Limited (https://csfrequency.com)",
   "license": "Apache-2.0",
